vsn-package {vsn}	R Documentation

vsn

Description

vsn

Details

The main function of the package is vsn2. Interesting for its applications are also predict and the wrapper function justvsn.

vsn2 can be applied to objects of class ExpressionSet, NChannelSet, AffyBatch (from the affy package) and RGList (from the limma package), matrix and vector. It returns an object of class vsn, which contains the results of fitting the vsn model to the data.

The most common use case is that you will want to construct a new data object with the vsn-normalized data whose class is the same as that of the input data and which preserves the metadata. This can be achieved by

    fit = vsn2(x, ...)
    nx = predict(fit, newdata=x)
  

To simplify this, there exists also a simple wrapper justvsn.

Author(s)

Wolfgang Huber

[Package vsn version 3.54.0 Index]