vsn {vsn}	R Documentation

Class to contain result of a vsn fit

Description

Class to contain result of a vsn fit

Creating Objects

new("vsn") vsn2(x) with x being an ExpressionSet.

Slots

coefficients:

A 3D array of size (number of strata) x (number of columns of the data matrix) x 2. It contains the fitted normalization parameters (see vignette).

strata:

A factor of length 0 or n. If its length is n, then its levels correspond to different normalization strata (see vignette).

mu:

A numeric vector of length n with the fitted parameters \hat{μ}_k, for k=1,...,n.

sigsq:

A numeric scalar, \hat{σ}^2.

hx:

A numeric matrix with 0 or n rows. If the number of rows is n, then hx contains the transformed data matrix.

lbfgsb:

An integer scalar containing the return code from the L-BFGS-B optimizer.

hoffset:

Numeric scalar, the overall offset c- see manual page of vsn2.

calib:

Character of length 1, see manual page of vsn2.

Methods

[

Subset

dim

Get dimensions of data matrix.

nrow

Get number of rows of data matrix.

ncol

Get number of columns of data matrix.

show

Print a summary of the object

exprs

Accessor to slot hx.

coef, coefficients

Accessors to slot coefficients.

Author(s)

Wolfgang Huber

See Also

vsn2

Examples

  data("kidney")
  v = vsn2(kidney)
  show(v)
  dim(v)
  v[1:10, ]

[Package vsn version 3.52.0 Index]