vsnInput {vsn}	R Documentation

Class to contain input data and parameters for vsn functions

Description

Class to contain input data and parameters for vsn functions

Creating Objects

new("vsnInput")

Slots

x:

A numeric matrix with the input data.

reference:

An object of vsn, typically this would have been obtained from a previous fit to a set of reference arrays (data).

strata:

A factor of length 0 or n. If its length is n, then its levels correspond to different normalization strata (see vsn2).

ordered:

Logical scalar; are the rows reordered so that the strata are contiguous.

lts.quantile:

Numeric scalar, seevsn2.

subsample:

Integer scalar, seevsn2.

verbose:

Logical scalar, seevsn2.

calib

Character of length 1, see manual page of vsn2.

pstart:

A 3D array of size (number of strata) x (number of columns of the data matrix) x 2. It contains the start parameters.

optimpar:

List with parameters for the numerical optimiser L-BFGS-B; see the manual page of vsn2.

Methods

[

Subset

dim

Get dimensions of data matrix.

nrow

Get number of rows of data matrix.

ncol

Get number of columns of data matrix.

show

Print a summary of the object

Author(s)

Wolfgang Huber

See Also

vsn2

[Package vsn version 3.52.0 Index]