justvsn {vsn}	R Documentation

Wrapper functions for vsn

Description

justvsn is equivalent to calling

  fit = vsn2(x, ...)
  nx = predict(fit, newdata=x, useDataInFit = TRUE)

vsnrma is a wrapper around vsn2 and rma.

Usage

justvsn(x, ...)
vsnrma(x, ...)

Arguments

x
For justvsn, any kind of object for which vsn2 methods exist. For vsnrma, an AffyBatch.

...
Further arguments that get passed on to vsn2.

Details

vsnrma does probe-wise background correction and between-array normalization by calling vsn2 on the perfect match (PM) values only. Probeset summaries are calculated with the medianpolish algorithm of rma.

Value

justvsn returns the vsn-normalised intensities in an object generally of the same class as its first argument (see the man page of predict for details). It preserves the metadata.

vsnrma returns an ExpressionSet.

Author(s)

Wolfgang Huber

See Also

vsn2

Examples

##--------------------------------------------------
## use "vsn2" to produce a "vsn" object
##--------------------------------------------------
data("kidney")
fit = vsn2(kidney)
nkid = predict(fit, newdata=kidney)

##--------------------------------------------------
## justvsn on ExpressionSet
##--------------------------------------------------
nkid2 = justvsn(kidney)
stopifnot(identical(exprs(nkid), exprs(nkid2)))

##--------------------------------------------------
## justvsn on RGList
##--------------------------------------------------
rg = new("RGList", list(R=exprs(kidney)[,1,drop=FALSE], G=exprs(kidney)[,2,drop=FALSE]))
erge = justvsn(rg)

[Package vsn version 3.52.0 Index]